[ci] Enhancement: Switch to new CI architecture
