--- .github/workflows/catloop_new.yml
@@ -1,99 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT license.
-
-name: Main Catloop (New)
-
-on:
-  workflow_run:
-    workflows: ["Main Catloop"]
-    types:
-      - completed
-
-env:
-  LIBOS: catloop
-  SERVER: ${{ secrets.CATMEM_HOSTNAME_B }}
-  CLIENT: ${{ secrets.CATMEM_HOSTNAME_B }}
-  SERVER_ADDR: 10.3.1.31
-  CLIENT_ADDR: 10.3.1.31
-
-jobs:
-
-  debug-pipeline:
-    name: Debug Pipeline
-    runs-on: ubuntu-latest
-    steps:
-    - name: Checkout
-      uses: actions/checkout@v3
-    - name: Setup SSH
-      shell: bash
-      run: |
-        mkdir -p $HOME/.ssh/
-        echo "${{ secrets.SSHKEY }}" > "$HOME/.ssh/id_rsa"
-        chmod 400 $HOME/.ssh/id_rsa
-        echo "Host *" > $HOME/.ssh/config
-        echo -e "\tStrictHostKeyChecking no" >> $HOME/.ssh/config
-        echo -e "\tIdentityFile $HOME/.ssh/id_rsa" >> $HOME/.ssh/config
-        echo -e "\tIdentitiesOnly yes" >> $HOME/.ssh/config
-        echo -e "\tPasswordAuthentication no" >> $HOME/.ssh/config
-        echo -e "\tUser ${{ secrets.USERNAME }}" >> $HOME/.ssh/config
-        echo -e "\tPort ${{ secrets.PORTNUM }}" >> $HOME/.ssh/config
-    - name: Run
-      run: |
-        python3 tools/demikernel_ci_new.py \
-          --server $SERVER \
-          --client $CLIENT \
-          --repository demikernel/demikernel \
-          --branch origin/${{ github.ref_name }} \
-          --libos $LIBOS \
-          --debug \
-          --test-unit --test-system all --delay 2 --enable-nfs \
-          --server-addr $SERVER_ADDR \
-          --client-addr $CLIENT_ADDR
-    - name: Archive Logs
-      if: always()
-      uses: actions/upload-artifact@v3
-      with:
-        name: debug-pipeline-logs
-        path: |
-          **/*.stdout.txt
-          **/*.stderr.txt
-
-  release-pipeline:
-    name: Release Pipeline
-    needs: debug-pipeline
-    runs-on: ubuntu-latest
-    steps:
-    - name: Checkout
-      uses: actions/checkout@v3
-    - name: Setup SSH
-      shell: bash
-      run: |
-        mkdir -p $HOME/.ssh/
-        echo "${{ secrets.SSHKEY }}" > "$HOME/.ssh/id_rsa"
-        chmod 400 $HOME/.ssh/id_rsa
-        echo "Host *" > $HOME/.ssh/config
-        echo -e "\tStrictHostKeyChecking no" >> $HOME/.ssh/config
-        echo -e "\tIdentityFile $HOME/.ssh/id_rsa" >> $HOME/.ssh/config
-        echo -e "\tIdentitiesOnly yes" >> $HOME/.ssh/config
-        echo -e "\tPasswordAuthentication no" >> $HOME/.ssh/config
-        echo -e "\tUser ${{ secrets.USERNAME }}" >> $HOME/.ssh/config
-        echo -e "\tPort ${{ secrets.PORTNUM }}" >> $HOME/.ssh/config
-    - name: Run
-      run: |
-        python3 tools/demikernel_ci_new.py \
-          --server $SERVER \
-          --client $CLIENT \
-          --repository demikernel/demikernel \
-          --branch origin/${{ github.ref_name }} \
-          --libos $LIBOS \
-          --test-unit --test-system all --delay 2 --enable-nfs \
-          --server-addr $SERVER_ADDR \
-          --client-addr $CLIENT_ADDR
-    - name: Archive Logs
-      if: always()
-      uses: actions/upload-artifact@v3
-      with:
-        name: release-pipeline-logs
-        path: |
-          **/*.stdout.txt
-          **/*.stderr.txt
