--- .github/workflows/catnip_new.yml
@@ -1,58 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT license.
-
-name: Main Catnip (New)
-
-on:
-  workflow_run:
-    workflows: ["Main Catnip"]
-    types:
-      - completed
-
-env:
-  LIBOS: catnip
-  SERVER: ${{ secrets.CATNIP_HOSTNAME_A }}
-  CLIENT: ${{ secrets.CATNIP_HOSTNAME_B }}
-  SERVER_ADDR: 10.3.1.50
-  CLIENT_ADDR: 10.3.1.51
-
-jobs:
-
-  release-pipeline:
-    name: Release Pipeline
-    runs-on: ubuntu-latest
-    steps:
-    - name: Checkout
-      uses: actions/checkout@v3
-    - name: Setup SSH
-      shell: bash
-      run: |
-        mkdir -p $HOME/.ssh/
-        echo "${{ secrets.SSHKEY }}" > "$HOME/.ssh/id_rsa"
-        chmod 400 $HOME/.ssh/id_rsa
-        echo "Host *" > $HOME/.ssh/config
-        echo -e "\tStrictHostKeyChecking no" >> $HOME/.ssh/config
-        echo -e "\tIdentityFile $HOME/.ssh/id_rsa" >> $HOME/.ssh/config
-        echo -e "\tIdentitiesOnly yes" >> $HOME/.ssh/config
-        echo -e "\tPasswordAuthentication no" >> $HOME/.ssh/config
-        echo -e "\tUser ${{ secrets.USERNAME }}" >> $HOME/.ssh/config
-        echo -e "\tPort ${{ secrets.PORTNUM }}" >> $HOME/.ssh/config
-    - name: Run
-      run: |
-        python3 tools/demikernel_ci_new.py \
-          --server $SERVER \
-          --client $CLIENT \
-          --repository demikernel/demikernel \
-          --branch origin/${{ github.ref_name }} \
-          --libos $LIBOS \
-          --test-unit --test-system all --delay 2 \
-          --server-addr $SERVER_ADDR \
-          --client-addr $CLIENT_ADDR
-    - name: Archive Logs
-      if: always()
-      uses: actions/upload-artifact@v3
-      with:
-        name: release-pipeline-logs
-        path: |
-          **/*.stdout.txt
-          **/*.stderr.txt
